# docs/swagger.yaml
basePath: /
definitions:
...
  ark_internal_httpx.ErrorBody:
    properties:
      error:
        type: string
    type: object
...
  internal_modules_relation.sendRequestByPhoneReq:
    properties:
      telephone:
        type: string
    required:
    - telephone
    type: object
  internal_modules_relation.sendRequestReq:
    properties:
      to_user_id:
        type: string
    required:
    - to_user_id
    type: object
host: localhost:8080
info:
  contact:
    name: ZooArk Team
  description: 'ZooArk 叫車平台後端 API。除了 POST /api/users、GET /health 和 GET /readyz 之外，所有
    endpoint 都需要 Firebase Auth ID Token（`Authorization: Bearer <token>`）。金額欄位以 TWD
    分（1/100 元）為單位。'
  title: ZooArk API
  version: 0.1.0
paths:
...
  /api/relations/friends:
    get:
      description: Returns the authenticated user's accepted friends.
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: List accepted friends
      tags:
      - Relation
  /api/relations/friends/{friend_id}:
    delete:
      description: Breaks the friendship between the authenticated user and `friend_id`
        (symmetric).
      parameters:
      - description: Friend user id
        in: path
        name: friend_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Remove friend
      tags:
      - Relation
  /api/relations/friends/{friend_id}/is:
    get:
      description: 'Returns whether `friend_id` is an accepted friend of the authenticated
        user. Response shape: `{"is_friend": true|false}`.'
      parameters:
      - description: Candidate user id
        in: path
        name: friend_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Check friendship
      tags:
      - Relation
  /api/relations/requests:
    post:
      consumes:
      - application/json
      description: Sends a friend request from the authenticated user to the user
        identified by `to_user_id`.
      parameters:
      - description: Target user id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_relation.sendRequestReq'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Send friend request
      tags:
      - Relation
  /api/relations/requests/{friend_id}:
    delete:
      description: Revokes a pending friend request the authenticated user previously
        sent to `friend_id`.
      parameters:
      - description: Target user id
        in: path
        name: friend_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Cancel outgoing friend request
      tags:
      - Relation
  /api/relations/requests/{friend_id}/accept:
    post:
      description: Accepts a pending friend request from `friend_id`. Both users become
        mutual friends.
      parameters:
      - description: Requester user id
        in: path
        name: friend_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Accept friend request
      tags:
      - Relation
  /api/relations/requests/{friend_id}/reject:
    post:
      description: Rejects a pending friend request from `friend_id`.
      parameters:
      - description: Requester user id
        in: path
        name: friend_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Reject friend request
      tags:
      - Relation
  /api/relations/requests/by-phone:
    post:
      consumes:
      - application/json
      description: Looks up a user by their registered phone number and sends a friend
        request from the authenticated user.
      parameters:
      - description: Target telephone
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_relation.sendRequestByPhoneReq'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Send friend request by phone
      tags:
      - Relation
  /api/relations/requests/received:
    get:
      description: Returns pending friend requests the authenticated user has received.
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: List received friend requests
      tags:
      - Relation
  /api/relations/requests/sent:
    get:
      description: Returns pending friend requests the authenticated user has sent.
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: List sent friend requests
      tags:
      - Relation
  /api/relations/search:
    get:
      description: Free-text user search by name or phone number. Returns a flat JSON
        array of matching user summaries.
      parameters:
      - description: Search query (name or phone)
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Search users
      tags:
      - Relation
...
schemes:
- http
- https
securityDefinitions:
  FirebaseAuth:
    description: Firebase ID Token, formatted as `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"